♻ refactor: actions 위치 수정
